\name{CopulaRegression-package}
\alias{CopulaRegression-package}
\alias{CopulaRegression}
\docType{package}
\title{
Bivariate copula-based regression models
}
\description{This R-packages presents a bivariate, copula-based model for the joint distribution of a pair of continuous and
discrete random variables. The two marginal random variables are modeled via generalized linear models, and their joint
distribution (represented by a parametric copula family) is estimated using maximum-likelihood techniques.
}
\details{
\tabular{ll}{
Package: \tab CopulaRegression\cr
Type: \tab Package\cr
Version: \tab 0.1-5\cr
Date: \tab 2014-09-04\cr
License: \tab GPL >=2\cr
}

}
\author{Nicole Kraemer <kraemer_r_packages@yahoo.de>
}
\references{N. Kraemer, E. Brechmann, D. Silvestrini, C. Czado (2013): Total loss estimation using copula-based regression models. Insurance: Mathematics and Economics 53 (3), 829 - 839.

}
\keyword{package}


